requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appdirs==1.4.4
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.4.0
beautifulsoup4==4.14.2
blender==1.4
#bpy==3.6.0
Brotli==1.1.0
#bs4==0.0.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
chromium==0.0.0
click==8.3.0
cssselect2==0.8.0
Cython==3.1.4
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.1
faiss-cpu==1.12.0
fastapi==0.120.0
ffmpy==0.6.4
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gpt4all==2.8.2
gradio==5.49.1
gradio_client==1.13.3
greenlet==3.2.4
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.0.1
langchain==1.0.2
langchain-classic==1.0.0
langchain-community==1.0.0a1
#langchain-huggingface==0.3.1
#langchain-mistralai==1.0.0
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.34
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mathutils==3.3.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
ollama==0.6.0
omegaconf==2.3.0
openai==2.3.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pdfkit==1.0.0
pexpect==4.9.0
pip==25.2
pillow==11.3.0
#playwright==1.55.0
prompt_toolkit==3.0.52
propcache==0.4.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydub==0.25.1
pydyf==0.11.0
#pyee==13
Pygments==2.19.2
pypdf==6.1.3
pyphen==0.17.2
#pyppeteer==2.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
ruff==0.14.2
safehttpx==0.1.7
safetensors==0.6.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.8.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.38.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zopfli==0.2.3.post1
zstandard==0.25.0